+++ app/src/main/res/layout/fragment_login.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_margin="20dp"
-    tools:context=".ui.view.auth.LoginFragment">
+    tools:context=".ui.view.auth.login.LoginFragment">
     <androidx.appcompat.widget.LinearLayoutCompat
         android:layout_marginHorizontal="20dp"
         android:layout_width="match_parent"
