+++ app/src/main/res/layout/fragment_register.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_margin="20dp"
-    tools:context=".ui.view.auth.RegisterFragment">
+    tools:context=".ui.view.auth.register.RegisterFragment">
     <androidx.appcompat.widget.LinearLayoutCompat
         android:layout_marginHorizontal="20dp"
         android:layout_width="match_parent"
