+++ app/src/main/res/values/strings.xml
@@ -21,6 +21,7 @@
     <string name="country_name_required">Country name is required</string>
     <string name="supplier_name_required">Supplier name is required</string>
     <string name="supplier_ruc_required">Supplier ruc is required</string>
+    <string name="supplier_ruc_too_short">Supplier ruc has 11 characters</string>
     <string name="category_required_error">Please select a category</string>
     <string name="country_required_error">Please select a country</string>
     <!-- TODO: Remove or change this placeholder text -->
